refactor(CH): reorganize mocks and test setup
* separate account and asset mocks
* move test storage setup closer to its actual usage
