+++ frame/clearing-house/src/mock/accounts.rs
@@ -1,0 +1,4 @@
+pub type AccountId = u64;
+
+pub const ADMIN: AccountId = 0;
+pub const ALICE: AccountId = 1;
