+++ frame/clearing-house/src/mock/assets.rs
@@ -1,0 +1,3 @@
+pub type AssetId = u128;
+pub const PICA: AssetId = 0;
+pub const USDC: AssetId = 1;
