+++ frame/clearing-house/src/mock/mod.rs
@@ -2,3 +2,9 @@
 pub mod governance_registry;
 #[cfg(test)]
 pub mod runtime;
+
+#[cfg(test)]
+pub mod accounts;
+
+#[cfg(test)]
+pub mod assets;
